this all is prep for login
