+++ Frontend/rasandbox/src/components/LoginWindow.vue
@@ -1,0 +1,10 @@
+<template>
+  <div>LOGIN</div>
+</template>
+<script lang="ts">
+import Vue from "vue";
+
+export default Vue.extend({
+  name: "LoginWindow",
+});
+</script>
