+++ Frontend/rasandbox/src/routes.ts
@@ -1,5 +1,5 @@
-import GameBox from "./components/GameBox.vue";
-import FilterModalVue from "./components/Filter-modal.vue";
+import LoginWindowVue from "./components/LoginWindow.vue";
+import Main from "./Main.vue";
 
 const routes = [
   {
@@ -9,7 +9,7 @@ const routes = [
   {
     path: "/games",
     name: "GameBox",
-    component: GameBox,
+    component: Main,
     props: (route: any) => ({
       page: route.query.page || 1,
       company: route.query.company || "",
@@ -21,9 +21,9 @@ const routes = [
     }),
   },
   {
-    path: "/filter", // Add a route for the filter page
-    name: "FilterModal",
-    component: FilterModalVue, // Make sure to import the FilterModal component
+    path: "/login",
+    name: "LoginWindow",
+    component: LoginWindowVue,
   },
 ];
